+++ src/ts/Core/Models/ABGroup.ts
@@ -71,7 +71,6 @@ export const HtmlEndcardTest = new DisabledABTest();
 export const ConsentUXTest = new ABTest(18, 19);
 
 // Load ABTests based on https://docs.google.com/spreadsheets/d/1-vtxnqIZ4FVusloKj4ZhaOtydSYNyk4BiQDI0Xny4qk/edit#gid=1252380516
-export const ZyngaOldApiHoldoutExperiment = new ABTest(13, 14);
 export const LoadExperiment = new ABTest(5, 6, 12, 16);
 export const LoadRefreshV4 = new ABTest(5, 6);
-export const ZyngaLoadRefreshV4 = new ABTest(5, 6, 12);
+export const ZyngaLoadRefreshV4 = new ABTest(5, 6, 12, 13, 14);